Fix clippy module inception (#108)
* clippy module inception: move kernel/kernel and kernel/mod into kernel

* clippy module inception: move actor/actor and actor/mod into actor

* module inception: move system/mod and system/system into system

* bonus clippy over tests
